bpf_map の fd が close されていなかったバグを修正

+++ pt-prio/common.h
@@ -4,7 +4,7 @@
 extern uint32_t repeat_cnt;
 extern uint32_t io_cnt;
 extern uint32_t disturb_num;
-extern double *io_start;
-extern double *io_end;
+extern unsigned long long *io_start;
+extern unsigned long long *io_end;
 
 #endif
